graphics added, some refactoring
Everything dice related (for the model, view, and controller) is in an
array instead of three separate variables.  This made it easier to
refactor part of the drawing into one function.

Images of the dice have replaced the three TextViews, and they are all
aligned to the center of the Roll button.